requirement.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-redis==5.0.0
djangorestframework==3.12.4
idna==3.2
pkg-resources==0.0.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
